add expect().toLooseDeepEqual() matcher
