+++ package.json
@@ -94,6 +94,7 @@
     "brfs": "^1.4.3",
     "browserify": "^13.0.0",
     "browserify-transform-tools": "^1.5.1",
+    "deep-equal": "^1.0.1",
     "ecstatic": "^1.4.0",
     "eslint": "^3.5.0",
     "falafel": "^1.2.0",
